feat: Move designTokens to the top level src directory
